requirements.txt
flask
flask_sqlalchemy
python-dotenv
pytest
werkzeug
psycopg2-binary
PyJWT 